+++ gems/aws-sdk-ec2/lib/aws-sdk-ec2/placement_group.rb
@@ -336,9 +336,6 @@ module Aws::EC2
     #   * `iam-instance-profile.id` - The instance profile associated with the
     #     instance. Specified as an ID.
     #
-    #   * `iam-instance-profile.name` - The instance profile associated with
-    #     the instance. Specified as an name.
-    #
     #   * `image-id` - The ID of the image used to launch the instance.
     #
     #   * `instance-id` - The ID of the instance.