+++ gems/aws-sdk-ecs/lib/aws-sdk-ecs/client.rb
@@ -832,6 +832,10 @@ module Aws::ECS
     #
     #  </note>
     #
+    # <note markdown="1"> Amazon Elastic Inference (EI) is no longer available to customers.
+    #
+    #  </note>
+    #
     # In addition to maintaining the desired count of tasks in your service,
     # you can optionally run your service behind one or more load balancers.
     # The load balancers distribute traffic across the tasks that are
@@ -931,15 +935,6 @@ module Aws::ECS
     # placement. For information about task placement and task placement
     # strategies, see [Amazon ECS task placement][7] in the *Amazon Elastic
     # Container Service Developer Guide*
-    #
-    # Starting April 15, 2023, Amazon Web Services will not onboard new
-    # customers to Amazon Elastic Inference (EI), and will help current
-    # customers migrate their workloads to options that offer better price
-    # and performance. After April 15, 2023, new customers will not be able
-    # to launch instances with Amazon EI accelerators in Amazon SageMaker,
-    # Amazon ECS, or Amazon EC2. However, customers who have used Amazon EI
-    # at least once during the past 30-day period are considered current
-    # customers and will be able to continue using the service.
     #
     #
     #
@@ -7325,6 +7320,10 @@ module Aws::ECS
     #
     #  </note>
     #
+    # <note markdown="1"> Amazon Elastic Inference (EI) is no longer available to customers.
+    #
+    #  </note>
+    #
     # You can allow Amazon ECS to place tasks for you, or you can customize
     # how Amazon ECS places tasks using placement constraints and placement
     # strategies. For more information, see [Scheduling Tasks][1] in the
@@ -7332,15 +7331,6 @@ module Aws::ECS
     #
     # Alternatively, you can use `StartTask` to use your own scheduler or
     # place tasks manually on specific container instances.
-    #
-    # Starting April 15, 2023, Amazon Web Services will not onboard new
-    # customers to Amazon Elastic Inference (EI), and will help current
-    # customers migrate their workloads to options that offer better price
-    # and performance. After April 15, 2023, new customers will not be able
-    # to launch instances with Amazon EI accelerators in Amazon SageMaker,
-    # Amazon ECS, or Amazon EC2. However, customers who have used Amazon EI
-    # at least once during the past 30-day period are considered current
-    # customers and will be able to continue using the service.
     #
     # You can attach Amazon EBS volumes to Amazon ECS tasks by configuring
     # the volume when creating or updating a service. For more infomation,
@@ -7909,14 +7899,9 @@ module Aws::ECS
     #
     #  </note>
     #
-    # Starting April 15, 2023, Amazon Web Services will not onboard new
-    # customers to Amazon Elastic Inference (EI), and will help current
-    # customers migrate their workloads to options that offer better price
-    # and performance. After April 15, 2023, new customers will not be able
-    # to launch instances with Amazon EI accelerators in Amazon SageMaker,
-    # Amazon ECS, or Amazon EC2. However, customers who have used Amazon EI
-    # at least once during the past 30-day period are considered current
-    # customers and will be able to continue using the service.
+    # <note markdown="1"> Amazon Elastic Inference (EI) is no longer available to customers.
+    #
+    #  </note>
     #
     # Alternatively, you can use`RunTask` to place tasks for you. For more
     # information, see [Scheduling Tasks][1] in the *Amazon Elastic
@@ -10446,7 +10431,7 @@ module Aws::ECS
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-ecs'
-      context[:gem_version] = '1.161.0'
+      context[:gem_version] = '1.162.0'
       Seahorse::Client::Request.new(handlers, context)
     end